update: module mkt

+++ src/app/core/local-storage/local-storeage-service.ts
@@ -1,42 +1,57 @@
 import { LocalStorageKey } from "./local-storage-key";
 import { UserModel } from "../../models/user-model";
-import { Injectable } from "@angular/core";
+import { inject, Injectable, PLATFORM_ID } from "@angular/core";
+import { isPlatformBrowser } from "@angular/common";
 
 @Injectable({
     providedIn: 'root'
 })
 export class LocalStorageSerice {
 
+    platformId = inject(PLATFORM_ID);
+
     clean(): void {
-        localStorage.clear();
+        if (isPlatformBrowser(this.platformId)) {
+            localStorage.clear();
+        }
     }
     getToken(): string {
-        return localStorage.getItem(LocalStorageKey.TOKEN) ?? '';
+        if (isPlatformBrowser(this.platformId)) {
+            return localStorage.getItem(LocalStorageKey.TOKEN) ?? '';
+        }
+        return '';
     }
     setToken(token: string) {
-        localStorage.setItem(LocalStorageKey.TOKEN, token);
+        if (isPlatformBrowser(this.platformId)) {
+            localStorage.setItem(LocalStorageKey.TOKEN, token);
+        }
     }
     removeToken(): void {
-        localStorage.removeItem(LocalStorageKey.TOKEN);
+        if (isPlatformBrowser(this.platformId)) {
+            localStorage.removeItem(LocalStorageKey.TOKEN);
+        }
     }
     getUserInfo(): UserModel {
-        return JSON.parse(localStorage.getItem(LocalStorageKey.USER_ID) ?? '');
+        let userModel: UserModel = {};
+        if (isPlatformBrowser(this.platformId)) {
+            if (localStorage.getItem(LocalStorageKey.USER_ID)) {
+                userModel = JSON.parse(localStorage.getItem(LocalStorageKey.USER_ID) ?? '');
+            }
+        }
+        return userModel;
     }
     setUserInfo(model?: UserModel): void {
-        localStorage.setItem(LocalStorageKey.USER_ID, model == undefined || model == null ? '' : JSON.stringify(model));
+        if (isPlatformBrowser(this.platformId)) {
+            localStorage.setItem(LocalStorageKey.USER_ID, model == undefined || model == null ? '' : JSON.stringify(model));
+        }
     }
     isLoggedIn(): boolean {
-        debugger;
-        var model = this.getUserInfo();
-        if (model) {
-            return true;
+        if (isPlatformBrowser(this.platformId)) {
+            var model = this.getUserInfo();
+            if (model) {
+                return true;
+            }
         }
         return false;
-    }
-    getScoinCode() {
-        return localStorage.getItem(LocalStorageKey.SCOIN_CODE) ?? '';
-    }
-    setScoinCode(code: string): void {
-        localStorage.setItem(LocalStorageKey.SCOIN_CODE, code);
     }
 }
